requirements.txt
python-docx>=1.2.0
openai>=1.12.0
tenacity>=8.2.0
tqdm>=4.66.0
click>=8.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# PDF support
pdf2docx>=0.5.8
PyPDF2>=3.0.0

# GUI support (optional)
tkinterdnd2>=0.3.0

# Performance monitoring (optional)
psutil>=5.9.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
supabase>=2.3.0
stripe>=7.0.0
